ethereum/RandomBeacon: Only generate one beacon per `set_beacon()` call.
The current Rust glue only will ever call i`set_beacon()` after the
epoch transition happens, so the code to generate the next epoch's
beacon is unused and unneeded.

This also has the potential to mess with the planned use of block number
by the registry to obtain global consistency.