Support owned dependency injection
This changes the type that can be returned by the component factory to
either Box<Arc<Trait>> or Box<Box<Trait>>. The selected injector then
determines whether the value from the factory should be converted into
a shared value or kept as an owned one.

In case a Box<Arc<Trait>> is returned (as was the only possibility up
till now) then only shared dependency injection is possible, while
returning a Box<Box<Trait>> supports both.