Make commit/reveal messages consensus-dependent
Previously the format of commitment and reveal messages was defined by
the consensus interface. Some backends have different requirements and
need to use their own format, so this commit changes the interface to
make the format of commitment and reveal messages opaque.

To handle signing it introduces a new ConsensusSigner interface which
provides two methods: sign_commitment and sign_reveal. The returned
structures can then be used to communicate with consensus.