Support per-method attributes in enclave RPC API
While the enclave RPC dispatcher already supported per-method attributes
like the requirement for client attestation, it was not possible to
configure these attributes through the API definitions.

This commit introduces method attributes using the standard Rust syntax
for attributes, starting with the client_attestation attribute. This
enables a specific method to not required client attestation while it is
required for other methods.