Handle empty root hash in trie
Previously the caller needed to handle the case where the root hash was
equal to the empty hash and pass None in this case. This introduced
subtle bugs where the caller forgot to do that and instead just copied
the root hash from the block data structure.

This commit changes the trie to automatically handle the case where the
root hash is equal to the empty hash and treats this case as if None was
passed as the root hash.