go/registry: Refactor the metrics insturmentation
Instead of doing it at the registry gRPC server layer, use a
`api.Backend` wrapper object applied automatically by the module level
ctor to handle metrics.

In the future it probably will be a good idea to move the registry
metrics over to using the pub/sub interface instead of intercepting
calls (in fact, this will be required once registry implementations
that actually expire node registrations are in use) as well.