Table generator iteration
Add ParamIter and framework for specifying the order in which
simulation parameter values are incremented (in odometric fashion),
when multiple iterations of simulation parameters are specified.

The simulator returns parallelization information as a value, so the
output of the simulator can eventually be HTML or Markdown tables.

SQUASH later commits into here.