go/tendermint/db/bolt: Only copy key/value on iterator accessors
The main benfit of this is avoiding the allocate/copy and associated GC
pressure due to copying the key (`bolt.fromBoltDBKey`) in the profiler
output.

The gain isn't as large as it could be due to the iavl library's liberal
use of the key and value accessors, but it's unfortunately unsafe not to
copy.

Part of #921.