averaging, reproducible simulations
This CL introduces averaging over a specified number of runs to get
better numbers (though we could probably get the same effect by
simulating more transactions).  It fixes a bug in the
LogicalShardingFilter which did not consistently assign replacement
shard number to the high-frequency locations due to set (map key)
iteration being inconsistent from run to run.  We introduce
ConsistentFind to LocationSet and now the simulator can produce
reproducible speedup numbers given the same random number seeds.

The default string value to iterflags will be a value that can be used
to set the flag, so flag.PrintDefault() will be more informative.
Some of the iterflag flags have values that will cause iteration.
This must be suppressed when using canned input via --input file flag,
but is probably good for new users of the tool.