Add cryptokitties distribution
The CryptoKitties distribution is half 3-party and half 2-party
transactions.  The 3-party transactions are breeding transactions,
involving the CryptoKitties contract, a dam kitty owner, and a sire
kitty owner.  The 2-party transactions are birthing transactions,
involving an "auto-birther" who invokes the "giveBirth" transaction on
the dam to create the new child kitten.