go/tendermint: Disable the transaction indexer
The transaction indexer was being ran with the default config and was
"only" indexing every transaction by hash.  Since we aren't using the
index, this might as well be disabled to save on a substantial amount
of disk space.

Note: The transaction index database on existing deployments will need
to be removed manually after a version with this commit is deployed.

Implements #1029.