registry: Add `GetRuntimes()`, rename old method to `WatchRuntimes()`
There is a lack of a method to poll the currently registered runtimes,
which is useful for writing tooling, so `GetRuntimes()` now exists in
the spirit of `GetNodes()`.  The existing method with the same name
is renamed to `WatchRuntimes()`.

WARNING: This breaks the roothash migration script, that the rest of the
branch is intended to replace.