go/roothash: Make the round timeout configurable
A configurable round timeout is beneficial especially while it is not
possible to interrupt a worker (#1010).  This parameter should be
globally consistent across all nodes (ie: be consensus backed, #965)
when there are multiple nodes running.

The parameter is `roothash.round_timeout`, with the default being the
old hard-coded value of 10 seconds.