go/scheduler/trivial: Fix the runtime list per epoch
This change fixes the runtime list on a per-epoch basis.  This should
allow changes to the runtime descriptor to correctly be factored into
scheduling decisions consistently across the entire network, as the
runtime descriptor used by each node is guaranteed to be consistent if
the appropriate registry backend is used.

A side-effect of this behavior is that the scheduler will no longer
optimistically try to elect a committee for newly registered runtimes
if sufficient nodes are available, as ALL runtime changes are applied
on the epoch transition.