go/storage: Add a simple proof-of-storage mechanism (aka receipts)
Till the v1 storage design with proper merkle proofs can be
implemented, add a simple proof-of-storage mechanism based around
an explicity query.

While this consumes an extra round-trip, it is significantly easier to
implement, and 1xRTT per batch is likely negligible in the grand scheme
of things.

WARNING: This is BREAKING and UNMIGRATABLE.