go/tendermint: Use the address book for the genesis validators
The `PersistentPeers` hack causes log spam and potential problems,
so replace it with a different hack consisting of forcing the genesis
validator addresses into the address book on startup.

The "correct" way to actually handle all of this is probably "seed nodes
with the validators setup as persistent peers", but that requires
running tendermint nodes that just seed.