runtime: Remove the old encryption interface
It will be nice if I could test the key manager again, and the old
interface wasn't implemented for the new tree, and it was evil.

This sticks a new (somewhat less evil) example encrypted mkvs
implementation in the simple-keyvalue runtime, under the assumption
that actual runtimes will implement their own idea of encrypted storage.