Increase SGX E2E tests parallelism to 3
Currently, we run 3 End-to-End tests on SGX. Since the SGX-enabled
Buildkite instances are not the fastest, this represents the slowest
part of the whole Buildkite testing pipeline.
The job that runs 2 End-to-End tests usually takes between 7 and 8
minutes, whereas the job that runs a single End-to-End test only takes
between 3 and 4 minutes.
In combination with increasing the number of SGX-enabled Buildkite
agents, we can increase SGX End-to-End tests parallelism to 3 and expect
overall build times to reduce for more than 3 minutes.