go/worker/{compute,merge}: Handle early discrepancy event
Previously, if a discrepancy event was seen before the batch (compute)
or results (merge) it was ignored, leading to timeouts.

This commit changes the states and processing so that the event is
instead recorded in StateWaitingForBatch/StateWaitingForResults and
batch/results processed immediately after being received in case of a
pending discrepancy event.