go/storage: Remove tree LRU cache
It does not make much sense to cache multiple trees on local backends
which are usually backed by memory-mapped files and thus do their own
caching.

For remote backends we now have a NodeDB that caches nodes so that
should be enough.

This also removes the --storage.root_cache.lru_size argument.