go/registry: Allow nodes to opt-in additional roles
Instead of requiring the registration to lapse for a node to opt-in
additional roles, always allow increasing the node roles.  Attempting to
remove any roles will still result in a rejection.

This still does not allow genesis validators to be compute workers
because going from no runtimes to having runtimes, gets rejected by the
registry.