go: Switch from `codec:` to `json:`
This is a design decision that I do not have strong feelings about.
Everywhere that already got this treatment are places where either the
data is only ever JSON, or is encoded/decoded both as JSON and CBOR.

Everything in this commit is where things will only ever be CBOR, so
this could instead use `cbor:` which the planned replacement also
accepts.

Note:
 * `toarray` annotations have not been changed and are left as is for
   now, since this is just laying the groundwork.