go/tendermint/apps/staking: Do not load all accounts to cache lookups
Previously staking.NewSnapshot fetched all accounts in order to populate
the cache. As there could be many accounts this doesn't seem the most
efficient thing to do.

Instead, this commit refactors the Snapshot to become StakeCache, a
lookup cache for stake queries. Escrow account balances are populated
on first access.