go/tendermint/apps/scheduler: Elect validators based on top-N staked
Top-n in this context is top-N out of the set of entities that actually
are running candidate validator nodes.  If someone happens to have the
most stake, but is not running a potential validator node at the time
of the election, their entity will not be included in the top-N staked.