go/tendermint/apps/scheduler: Change validator elections again
Meetings happened when I was asleep, so change the algorithm to a
sensible version of what was decided on.

Validators are now picked, 1 per entity (at random if multiple candiates
ran by a single entity are available), in order of decending stake, up
to the maximum number of validators.

Unlike the previous version of this where eligibility was also
restricted to the top-N entities by stake, this election algorithm will
continue to go down the entity list if required in an attempt to fill
out the validator set.

Incentives are now "any entity that is running a validator at the time
of the election, that meets the minimum stake threshold", rather than
being restricted to "top-N" as well.

Note: The "1 per entity" limit is configurable.  If more are allowed,
then each entity (in descending order) will have the maximum number of
nodes that they have the stake to support be elected.  This isn't great
but the limit being higher should only ever happen in test environments
anyway.