go/consensus: Use common transaction format
Previously each of the Oasis components (registry, roothash, staking,
etc.) defined its own internal transaction format when translating
method calls to the consensus layer.

After we realized that we need to add common things like gas fees and
nonces to most of the method calls, a common transaction format results
in less duplication and allows for things like common transaction
authentication handlers.

This commit changes the following:

- Oasis component APIs are changed so that they only perform read-only
  operations. For mutation they expose functions for generating
  transactions which can be submitted to the consensus layer via a new
  SubmitTx method.

- The CLI is changed so that there is a single way of submitting
  transactions. Existing subcommands (registry, staking) only generate
  signed transactions which is similar to before but now using the
  common transaction format.

- Together with the common transaction format there is now a common way
  to register method names and for Tendermint applications to specify
  which methods they support.

- The consensus backend now has a notion of a transaction auth handler
  which performs things like nonce and fee checking. The handler is
  invoked for each transaction made possible by the common transaction
  format.

- Some consensus-related cleanup.

Due to changing the transaction format this is a BREAKING change.