go/consensus: Watch transaction recheck failures in SubmitTx
A transaction could be valid at the time of submission but become
invalid later. Previously SubmitTx never noticed that a transaction
became invalid and just kept waiting for it to be included in a block
which of course can never happen for invalid transactions.

Since Tendermint's mempool doesn't support such notifications, this adds
support for subscribing to when a transaction becomes invalid during a
mempool re-check (which is triggered by the mempool on each block).