Make: Make Punch version file creation lazy and add a clean target
Defer creation of a temporary Punch file to the first time it is used.

Add clean-version-files target that can be used to manually clean up
temporary Punch version files.
Note that these files are created in the /tmp directory which is
normally mounted as a tmpfs file system meaning these files are kept in
memory and will be automatically gone after the system is restarted.