go/oasis-node/cmd/debug/consim: Add `--consim.memdb`
While it's possible to use the pruner to try to prevent the ABCI induced
disk usage from growing out of control, having to save/prune is rather
slow, even when backed by tmpfs.

Enabling this option will keep ABCI state in memory, dramatically
speeding up cases where only the final dump is required, especially if
a sensible configuration is used for the pruner.