go/runtime/committee: Restore previously picked node in RR selection
Previously the round-robin node selection policy would randomize the order on
every update ignoring the currently picked node. This would cause the current
node to flip on each update causing problems with EnclaveRPC which is stateful.

The fix makes the round-robin node selection policy attempt to restore the
currently picked node on each update. This means that in case the node is still
in the node list, it will not change.