runtime: Make CheckTx result more generic
The previous definition made it hard for the runtime host to reason about the
returned status code of CheckTx unless it was tied to the result format
defined by the runtime.

This is a breaking change to the Runtime Host Protocol as the
RuntimeCheckTxBatchResponse structure has changed. It also breaks the
transaction::Dispatcher API.