go/consensus/tendermint: Only close DBs in archive node mode
Previously the Tendermint-related state stores were always closed during
Stop in the common node. But in case the node is a full node, the stores
are owned by the Tendermint Core Node and as such should not be closed
to avoid a panic during queries.

In addition, the more correct behavior is to close the stores during
service cleanup to ensure that all other services that may use them have
already been stopped.