go/runtime: Fix configuring the SGX-based provisioner
Previously there was an edge case when the SGX loader was configured
and remapping to non-SGX was forced (e.g. on client nodes). This would
result in an invalid SGX configuration that resulted in a strange error
message about a missing SIGSTRUCT.

This has now been changed so that remapping happens independent of
whether an SGX loader is configured or not.